Fix building on CY2025 on MacOS (#1031)
### Fix building with MacOS on CY2025

### Summarize your change.

The versions of openexr and png included with CY2025 introduce rpath
setups, which ends up conflicting with RV's own rpath management on
macos. And resulting in build errors like the following when RV tries to
overwrite them.
```
/Users/nelsonr/git/rv/_build_debug/RV_DEPS_OPENEXR/install/lib/libOpenEXRUtil-3_2_d.32.3.3.6.dylib (for architecture arm64) option "-add_rpath @loader_path/../lib" would duplicate path, file already has LC_RPATH for: @loader_path/../lib
```

The simplest olution is just to disable the rpath generation in those
libraries, and let RV handle it as its currently doing for earlier
versions. An corollary fix was required in the RV rpath script that
removes rpaths to not to fail if there is no rpath (since we now
effectively remove it at the build stage). Another script later in the
build process will add the correct rpath back in.

### Describe the reason for the change.

Fix building on MacOS with CY2025

### Describe what you have tested and on which operating system.

MacOS 26.1

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>